Added support for several bases in worker.
Minor changes:
Minor refactoring of model class.
Added virtual destructor in Scope class.
Added assignment operator in basis class.

+++ tests/test.cpp
@@ -5,7 +5,7 @@ int main() {
 
   Basis b;
   std::vector<vec3> bp(4);
-  std::vector<double> model_vertices(9);
+  std::vector<float> model_vertices(9);
 
   for (int i = 0; i < 4; ++i) {
     bp[i].x = i;
